[editor] Improve ai loading with cache
