+++ src/component/editor/ai-view.vue
@@ -202,6 +202,7 @@
 		private hoverData: any = null
 		private ctrl: boolean = false
 		private CodeMirrorLines!: HTMLElement
+		private jumpToLine: number | null = null
 
 		created() {
 			this.id = this.ai.id
@@ -292,6 +293,13 @@
 
 				this.show()
 
+				if (this.jumpToLine) {
+					setTimeout(() => {
+						this.scrollToLine(this.jumpToLine!)
+						this.jumpToLine = null
+					})
+				}
+
 				this.CodeMirrorLines = codeMirrorElement.querySelector('.CodeMirror-lines') as HTMLElement
 
 				// Lock scroll down
@@ -345,17 +353,16 @@
 			this.ai.modified = true
 			this.removeAllErrors()
 		}
+
 		public show() {
 			if (this.loaded) {
 				this.editor.refresh()
 			} else {
 				this.loading = true
-				LeekWars.get('ai/get/' + this.id).then(data => {
-					Vue.set(this.ai, 'code', data.ai.code)
+				fileSystem.load(this.ai).then(() => {
 					this.editor.setValue(this.ai.code)
 					this.editor.getDoc().clearHistory()
 					this.editor.refresh()
-					this.analyzeV1()
 					this.loaded = true
 					this.loading = false
 					setTimeout(() => this.editor.refresh())
@@ -446,14 +453,13 @@
 		}
 
 		setAnalyzerTimeout() {
-			// this.analyze()
 			clearTimeout(this.analyzerTimeout)
 			this.analyzerTimeout = setTimeout(() => {
-				this.analyzeV1()
-			}, 500)
+				this.ai.analyze()
+			}, 1000)
 		}
 
-		analyze() {
+		analyzeV2() {
 			const content = this.editor.getValue()
 			return LeekWars.analyzer.analyze(this.ai, this.editor.getDoc().getValue()).then((problems) => {
 				this.$emit('problems', problems)
@@ -464,28 +470,6 @@
 					Vue.delete(this.errorOverlays, entrypoint)
 				}
 			})
-		}
-
-		public analyzeV1() {
-
-			this.updateIncludes()
-
-			// Search /* */ comments first
-			let match
-			const code = this.editor.getValue()
-			const comments: {[key: number]: string} = {}
-			const comment_regex = /\/\*([^]*?)\*\/\s*/gm
-			while ((match = comment_regex.exec(code)) != null) {
-				const content = match[1].trim().split("\n").map(line => line.replace(/^\s*\*\s?/, '')).join("\n").trim()
-				comments[match.index + match[0].length] = content
-			}
-			this.ai.comments = comments
-
-			// console.log("Comments", comments)
-
-			this.updateFunctions()
-			this.updateClasses()
-			this.updateGlobalVars()
 		}
 
 		public change(CodeMirror: any, changes: CodeMirror.EditorChange) {
@@ -554,12 +538,14 @@
 				}
 			}
 		}
+
 		public formatCode() {
 			import(/* webpackChunkName: "js-beautify" */ "js-beautify").then(js_beautify => {
 				console.log(js_beautify)
 				this.editor.setValue(js_beautify.js_beautify(this.editor.getValue(), {indent_size: 1, indent_char: '\t'}))
 			})
 		}
+
 		public commentCode() {
 			const start = this.document.getCursor('from').line
 			const end = this.document.getCursor('to').line
@@ -594,6 +580,7 @@
 				}
 			}
 		}
+
 		public getLineIndentation(lineNo: number) {
 			const line = this.document.getLine(lineNo)
 			let indent = ""
@@ -606,6 +593,7 @@
 			}
 			return indent
 		}
+
 		public unindentCode() {
 			const start = this.document.getCursor("start").line
 			const end = this.document.getCursor("end").line
@@ -616,6 +604,7 @@
 				}
 			}
 		}
+
 		public removeLine() {
 			const line = this.document.getCursor("start").line
 			if (this.document.somethingSelected()) {
@@ -625,6 +614,7 @@
 			}
 			this.document.setCursor(line)
 		}
+
 		public duplicateLine() {
 			if (this.document.somethingSelected()) {
 				const start = this.document.getCursor("start")
@@ -1326,331 +1316,16 @@
 			this.hintDialog = false
 			this.editor.removeKeyMap(this.dialogKeyMap)
 		}
+
 		public scrollToLine(line: number) {
-			this.document.setCursor({line, ch: 0})
-			const height = this.editor.getScrollInfo().clientHeight
-			const coords = this.editor.charCoords({line, ch: 0}, "local")
-			this.editor.scrollTo(null, (coords.top + coords.bottom - height) / 2)
-		}
-
-		public updateIncludes() {
-			// console.log("Update includes", this.ai.name)
-			// console.time("inc")
-			this.ai.includes = []
-			const code = this.document.getValue() || this.ai.code
-			const regex = /include\s*\(\s*["'](.*?)["']\s*\)/gm
-			let m
-			while (m = regex.exec(code)) {
-				const path = m[1]
-				const included = fileSystem.find(path, this.ai.folder)
-				if (included) {
-					// console.log("Found included", path, this.ai.folder, included)
-					this.ai.includes.push(included)
-					this.$emit("load", included)
-					LeekWars.analyzer.register(included).catch(e => {
-						// nothing
-					})
-				} else {
-					// console.warn("Included not found", path, this.ai.folder, included)
-				}
-			}
-			for (const entrypoint_id of this.ai.entrypoints) {
-				const entrypoint = fileSystem.ais[entrypoint_id]
-				if (entrypoint) {
-					this.$emit("load", entrypoint)
-				}
-			}
-			// console.timeEnd("inc")
-		}
-
-		public updateFunctions() {
-			const code = this.editor.getValue()
-			this.ai.functions = []
-			let match
-
-			const regex = /function\s+(\w+)\s*\(([^]*?)\)\s*{/gm
-			// Match [ full_match, javadoc, nom, arguments ]
-
-			while ((match = regex.exec(code)) != null) {
-
-				const line = code.substring(0, match.index).split("\n").length
-				let args = match[2].split(",")
-				if (args.length === 1 && args[0].trim() === '') { args = [] }
-				for (const a in args) {
-					args[a] = args[a].trim()
-					if (args[a].startsWith('@')) {
-						args[a] = args[a].substring(1)
-					}
-				}
-				let fullName = match[1] + "(" + args.join(", ") + ")"
-				let description = "<h4>" + i18n.t('leekscript.function_f', [fullName]) + "</h4><br>"
-				description += i18n.t('leekscript.defined_in', [this.ai.name, line])
-
-				const comment = this.ai.comments[match.index]
-				const javadoc = {
-					name: fullName,
-					description: "",
-					items: [] as any[]
-				}
-				// Add arguments from signature
-				for (const arg of args) {
-					javadoc.items.push({ type: 'param', name: arg, text: null})
-				}
-				// console.log(javadoc.items)
-				if (comment) {
-					const javadoc_lines = comment.split("\n")
-					const javadoc_regex = /^\s*@(\w+)(?:\s+([a-zA-Z_\u00C0-\u024F\u1E00-\u1EFF]+)\s*:?\s*)?(?:\s*:\s*)?(.*)$/
-					let match_javadoc
-					for (const jline of javadoc_lines) {
-						if (match_javadoc = javadoc_regex.exec(jline)) {
-							// console.log(match_javadoc)
-							const type = match_javadoc[1]
-							let name = match_javadoc[2]
-							let text = match_javadoc[3]
-							if (type === 'return') {
-								fullName += ' : ' + text
-							} else if (type === 'param') {
-								if (name) {
-									name = name.trim()
-									if (name.startsWith('@')) {
-										name = name.substring(1)
-									}
-								}
-								text = text.trim()
-								if (text.startsWith('@')) {
-									text = text.trim().substring(1)
-								}
-								if (args.includes(name) || args.includes(text)) {
-									// console.log('arg', name, text)
-									const existing = javadoc.items.find(i => i.type === 'param' && ((name.length && i.name === name) || (text.length && i.name === text)))
-									// console.log('existing', existing)
-									// existing.name = existing.text
-									existing.text = text
-									continue
-								}
-							}
-							javadoc.items.push({ type, name, text })
-						} else {
-							// const star = jline.indexOf("*")
-							// let formatted_line = jline.substring(star + 2)
-							// if (l === javadoc_lines.length - 1) {
-							// 	formatted_line = formatted_line.trim()
-							// }
-							if (jline.length) {
-								if (javadoc.description.length) {
-									javadoc.description += "\n"
-								}
-								javadoc.description += jline
-							}
-						}
-					}
-					// console.log("javadoc", javadoc)
-				}
-
-				// Escape
-				javadoc.description = LeekWars.protect(javadoc.description)
-				for (const item of javadoc.items) {
-					item.name = item.name ? LeekWars.protect(item.name) : item.name
-					item.text = item.text ? LeekWars.protect(item.text) : item.text
-				}
-
-				const fun = {
-					name: match[1],
-					fullName,
-					details: description,
-					type: 'user-function',
-					argumentCount: args.length,
-					arguments: args,
-					ai: this.ai,
-					line,
-					javadoc,
-					category: 4
-				}
-				// console.log(fun)
-				this.ai.functions.push(fun)
-			}
-		}
-
-		public updateClasses() {
-			const code = this.editor.getValue()
-			this.ai.classes = {}
-			let match
-
-			// Search classes
-			const class_regex = /class\s+(\w+)\s*(extends|{)/gm
-			while ((match = class_regex.exec(code)) != null) {
-				const line = code.substring(0, match.index).split("\n").length
-				const name = match[1]
-				const comment = this.ai.comments[match.index]
-				const javadoc = { name, description: comment, items: [] }
-				this.ai.classes[name] = {
-					name,
-					fullName: name,
-					details: "Classe <b>" + name + "</b>",
-					type: 'class',
-					ai: this.ai,
-					line,
-					category: 9,
-					javadoc,
-					fields: [],
-					static_fields: [
-
-					],
-					methods: [],
-					static_methods: []
-				}
-				if (this.ai.version >= 3) {
-					this.ai.classes[name].static_fields.push(
-						{ name: "name", fullName: "name", type: "static-field", category: 1, details: this.$t('leekscript.class_name') },
-						{ name: "super", fullName: "super", type: "static-field", category: 1, details: this.$t('leekscript.class_super') },
-						{ name: "fields", fullName: "fields", type: "static-field", category: 1, details: this.$t('leekscript.class_fields') },
-						{ name: "staticFields", fullName: "staticFields", type: "static-field", category: 1, details: this.$t('leekscript.class_staticFields') },
-						{ name: "methods", fullName: "methods", type: "static-field", category: 1, details: this.$t('leekscript.class_methods') },
-						{ name: "staticMethods", fullName: "staticMethods", type: "static-field", category: 1, details: this.$t('leekscript.class_staticMethods') },
-					)
-				}
-			}
-			// console.log("Classes", this.ai.classes)
-
-			// Search methods
-			const method_regex = /(?:public\s+)?(?:(static)\s+)?(\w+)\s*\(([\w\s,]*)\)\s*{/gm
-			while ((match = method_regex.exec(code)) != null) {
-
-				const name = match[2]
-				if (name === 'function' || name === 'for' || name === 'while' || name === 'if') continue
-
-				const is_static = !!match[1]
-				const line = code.substring(0, match.index).split("\n").length
-				let args = match[3].split(",")
-				if (args.length === 1 && args[0].trim() === '') { args = [] }
-				for (const a in args) {
-					args[a] = args[a].trim()
-					if (args[a].startsWith('@')) {
-						args[a] = args[a].substring(1)
-					}
-				}
-				let fullName = match[2] + "(" + args.join(", ") + ")"
-				let description = "<h4>" + i18n.t('leekscript.function_f', [fullName]) + "</h4><br>"
-				description += i18n.t('leekscript.defined_in', [this.ai.name, line])
-
-				const comment = this.ai.comments[match.index]
-				// console.log("comment", comment)
-				const javadoc = {
-					name: fullName,
-					description: "",
-					items: [] as any[]
-				}
-				// Add arguments from signature
-				for (const arg of args) {
-					javadoc.items.push({ type: 'param', name: arg, text: null})
-				}
-				// console.log(javadoc.items)
-				if (comment) {
-					const javadoc_lines = comment.split("\n")
-					const javadoc_regex = /^\s*@(\w+)(?:\s+([a-zA-Z_\u00C0-\u024F\u1E00-\u1EFF]+)\s*:?\s*)?(?:\s*:\s*)?(.*)$/
-					let match_javadoc
-					for (const jline of javadoc_lines) {
-						if (match_javadoc = javadoc_regex.exec(jline)) {
-							// console.log(match_javadoc)
-							const type = match_javadoc[1]
-							let name = match_javadoc[2]
-							let text = match_javadoc[3]
-							if (type === 'return') {
-								fullName += ' : ' + text
-							} else if (type === 'param') {
-								if (name) {
-									name = name.trim()
-									if (name.startsWith('@')) {
-										name = name.substring(1)
-									}
-								}
-								text = text.trim()
-								if (text.startsWith('@')) {
-									text = text.trim().substring(1)
-								}
-								if (args.includes(name) || args.includes(text)) {
-									// console.log('arg', name, text)
-									const existing = javadoc.items.find(i => i.type === 'param' && ((name.length && i.name === name) || (text.length && i.name === text)))
-									// console.log('existing', existing)
-									// existing.name = existing.text
-									existing.text = text
-									continue
-								}
-							}
-							javadoc.items.push({ type, name, text })
-						} else {
-							if (jline.length) {
-								if (javadoc.description.length) {
-									javadoc.description += "\n"
-								}
-								javadoc.description += jline
-							}
-						}
-					}
-					// console.log("javadoc", javadoc)
-				}
-
-				// Escape
-				javadoc.description = LeekWars.protect(javadoc.description)
-				for (const item of javadoc.items) {
-					item.name = item.name ? LeekWars.protect(item.name) : item.name
-					item.text = item.text ? LeekWars.protect(item.text) : item.text
-				}
-
-				// Find class
-				let clazz = null
-				for (const c in this.ai.classes) {
-					if (this.ai.classes[c].line! > line) break
-					clazz = this.ai.classes[c]
-				}
-				if (clazz) {
-					const method = {
-						name: match[2],
-						fullName,
-						details: description,
-						type: is_static ? 'user-static-method' : 'user-method',
-						argumentCount: args.length,
-						arguments: args,
-						ai: this.ai,
-						line,
-						javadoc,
-						category: 4
-					}
-					// console.log(fun)
-					if (is_static) {
-						clazz.static_methods.push(method)
-					} else {
-						clazz.methods.push(method)
-					}
-				}
-			}
-
-			// console.log("classes " + this.ai.name, this.ai.classes)
-		}
-
-		public updateGlobalVars() {
-
-			const code = this.editor.getValue()
-			this.ai.globals = {}
-			let match
-
-			// Search global vars
-			const global_regex = /global\s+(\w+)/gm
-			while ((match = global_regex.exec(code)) != null) {
-				const line = code.substring(0, match.index).split("\n").length
-				const name = match[1]
-				const comment = this.ai.comments[match.index]
-				const javadoc = { name, description: comment, items: [] }
-				this.ai.globals[name] = {
-					name,
-					fullName: name,
-					details: "Variable <b>" + name + "</b>",
-					type: 'variable',
-					ai: this.ai,
-					line,
-					category: 8,
-					javadoc,
-				}
+			// console.log("scrollToLine", line, this.document, this.editor)
+			if (this.document) {
+				this.document.setCursor({line, ch: 0})
+				const height = this.editor.getScrollInfo().clientHeight
+				const coords = this.editor.charCoords({line, ch: 0}, "local")
+				this.editor.scrollTo(null, (coords.top + coords.bottom - height) / 2)
+			} else {
+				this.jumpToLine = line
 			}
 		}
 
@@ -1669,6 +1344,7 @@
 				this.closeSearch()
 			}
 		}
+
 		public closeSearch() {
 			this.searchEnabled = false
 			if (this.searchOverlay) {
@@ -1676,6 +1352,7 @@
 				this.searchOverlay = null
 			}
 		}
+
 		@Watch('searchQuery')
 		public searchUpdate() {
 			const query = this.searchQuery.toLowerCase()
